--- Apresentacao_do_projeto/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/home/fill/anaconda3/envs/landmarks/bin/python"
-}
