--- Códigos/DeepFace.py
@@ -1,1169 +1,0 @@
-from keras.preprocessing import image
-import warnings
-warnings.filterwarnings("ignore")
-import time
-import os
-from os import path
-from pathlib import Path
-import gdown
-import numpy as np
-import pandas as pd
-from tqdm import tqdm
-import json
-import cv2
-from keras import backend as K
-import keras
-import tensorflow as tf
-import pickle
-import matplotlib.pyplot as plt
-
-from deepface import DeepFace
-from deepface.basemodels import VGGFace, OpenFace, Facenet, FbDeepFace, DeepID
-from deepface.extendedmodels import Age, Gender, Race, Emotion
-from deepface.commons import functions, realtime, distance as dst
-
-def make_dict(dataset_path):
-	dic = {}
-	os.chdir(dataset_path)   #Aqui entramos no diretório do dataset
-	class_folders = os.listdir('.') #Pegamos os strings referente aos diretórios referentes a cada classe
-	for class_folder in class_folders: #Loop para varrer cada diretório
-		i = 1 #Variável auxiliar para nomear as imagens
-		os.chdir(class_folder) #Entramos no diretório de uma classe
-		images = os.listdir('.') #Pegamos os strings de cada arquivo de imagem
-		for image in images: #Loop para varrer as imagens
-			#os.rename(image, class_folder+'.{}'.format(i)+'.jpg') #Renomeamos cada imagem com o nome no formato que queremos
-			dic['{}.{}'.format(class_folder, i)] = dataset_path+ chr(92) +class_folder+chr(92)+image
-			i = i+1 #Atualizamos a variável
-		os.chdir('..') #Após renomear todas as imagens em um diretório, saimos dele para ir ao próximo
-	os.chdir('..')
-	return dic
-
-def save_hash(img_dict, model_name = 'Ensemble', model = None, enforce_detection = True):
-
-	tic = time.time()
-
-	img_list = list(img_dict.keys())
-
-	if model_name == 'Ensemble':
-		print("Ensemble learning enabled")
-		
-		import lightgbm as lgb #lightgbm==2.3.1
-		
-		if model == None:
-			model = {}
-			
-			model_pbar = tqdm(range(0, 4), desc='Face recognition models')
-			
-			for index in model_pbar:
-				
-				if index == 0:
-					model_pbar.set_description("Loading VGG-Face")
-					model["VGG-Face"] = VGGFace.loadModel()
-				elif index == 1:
-					model_pbar.set_description("Loading Google FaceNet")
-					model["Facenet"] = Facenet.loadModel()
-				elif index == 2:
-					model_pbar.set_description("Loading OpenFace")
-					model["OpenFace"] = OpenFace.loadModel()
-				elif index == 3:
-					model_pbar.set_description("Loading Facebook DeepFace")
-					model["DeepFace"] = FbDeepFace.loadModel()
-		
-		#--------------------------
-		
-		model_names = ["VGG-Face", "Facenet", "OpenFace", "DeepFace"]
-		
-		pbar = tqdm(range(0,len(img_list)), desc='Avaliando Hash')
-		
-		df = pd.DataFrame(np.zeros([len(img_list), len(model_names)]),
-                  		  columns = model_names, index = list(img_dict.keys()))
-
-		df = df.astype(object)
-
-		representation = []
-
-		erro_keys = []
-		
-		#for instance in img_list:
-		for index in pbar:
-
-			erro = False
-
-			img1_key = img_list[index]
-			
-			representation.clear()
-			
-			for i in  model_names:
-				custom_model = model[i]
-
-				#input_shape = custom_model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some 						people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-
-				input_shape = custom_model.layers[0].input_shape
-				
-				if type(input_shape) == list:
-					input_shape = input_shape[0][1:3]
-				else:
-					input_shape = input_shape[1:3]
-				
-				try:
-					img1 = functions.detectFace(img_dict[img1_key], input_shape, enforce_detection = enforce_detection)
-				except:
-					erro = True
-					erro_keys.append(img1_key)
-					break
-
-				img1_representation = custom_model.predict(img1)[0,:]
-
-				representation.append(img1_representation)
-
-			if erro:
-				continue
-			
-			df.loc[list(img_dict.keys())[index]] = representation
-			
-		for i in erro_keys:
-			df.drop(index = i, inplace = True)				
-
-		return df
-	#------------------------------
-
-def metrics_dataframe(img_df, one_compair = False):
-	
-	cosine_distance = []
-	euclidean_distance = []
-	l2_distance = []
-	distances = []
-
-	metrics = ['Cosine', 'Euclidean', 'L2']
-	model = list(img_df.columns)
-
-	column_names = []
-	index_names = []
-	for i in model:
-		for j in metrics:
-			column_names.append(j+' with '+i)
-
-	for i in img_df.index:
-		for j in img_df.index:
-			index_names.append(i+ '-' +j)
-	
-	metrics_df = pd.DataFrame(columns = column_names, index = index_names)
-
-	metrics_df = metrics_df.astype(object)
-
-	if one_compair:
-
-		people_pbar = tqdm(range(1, len(img_df.index)), desc = 'Avaliando a primeira pessoa em relação à todas as outras')
-
-		for index in people_pbar:
-
-			image_index = img_df.index[0]
-			real_index = img_df.index[index]
-
-			cosine_distance.clear()
-			euclidean_distance.clear()
-			l2_distance.clear()
-			distances.clear()
-
-			for model in img_df.columns:
-			
-				#Cosine:
-				cosine_distance.append(dst.findCosineDistance(img_df.at[image_index, model], img_df.at[real_index, model]))
-
-				#Euclidean:
-				euclidean_distance.append(dst.findEuclideanDistance(img_df.at[image_index, model], img_df.at[real_index, model]))
-
-				#L2:
-				l2_distance.append(dst.findEuclideanDistance(dst.l2_normalize(img_df.at[image_index, model]), 										     dst.l2_normalize(img_df.at[real_index, model])))
-
-			for i in range(0, len(img_df.columns)):
-				distances.append(cosine_distance[i])
-				distances.append(euclidean_distance[i])
-				distances.append(l2_distance[i])
-
-			string = real_index + '-' + image_index
-
-			metrics_df.loc[string] = distances
-
-
-	else:
-		people_pbar = tqdm(range(0, len(img_df.index)), desc = 'Avaliando cada pessoa em relação à todas as outras')
-
-		for index in people_pbar:
-			
-			real_index = img_df.index[index]
-
-			for jndex in people_pbar:
-
-				real_jndex = img_df.index[jndex]
-
-				cosine_distance.clear()
-				euclidean_distance.clear()
-				l2_distance.clear()
-				distances.clear()
-				
-				for model in img_df.columns:
-				
-					#Cosine:
-					cosine_distance.append(dst.findCosineDistance(img_df.at[real_index, model], img_df.at[real_jndex, model]))
-
-					#Euclidean:
-					euclidean_distance.append(dst.findEuclideanDistance(img_df.at[real_index, model], img_df.at[real_jndex,
-model]))
-
-					#L2:
-					l2_distance.append(dst.findEuclideanDistance(dst.l2_normalize(img_df.at[real_index, model]), 										     dst.l2_normalize(img_df.at[real_jndex, model])))
-
-				for i in range(0, len(img_df.columns)):
-					distances.append(cosine_distance[i])
-					distances.append(euclidean_distance[i])
-					distances.append(l2_distance[i])
-
-				string = real_index + '-' + real_jndex
-
-				metrics_df.loc[string] = distances
-	
-	metrics_df.dropna(inplace = True)
-
-	return metrics_df
-
-	#------------------------------
-
-def ensemble_dataframe(metrics_df):
-
-	import lightgbm as lgb #lightgbm==2.3.1 
-
-	columns_names = ['Verified', 'Reality', 'Score']
-
-	metrics_df = metrics_df.drop('Euclidean with OpenFace', axis = 1)
-
-	verification_df = pd.DataFrame(columns = columns_names, index = metrics_df.index)
-	
-	#-------------------------------
-	#find deepface path
-	
-	home = str(Path.home())
-	
-	if os.path.isfile(home+'/.deepface/weights/face-recognition-ensemble-model.txt') != True:
-		print("face-recognition-ensemble-model.txt will be downloaded...")
-		url = 'https://raw.githubusercontent.com/serengil/deepface/master/deepface/models/face-recognition-ensemble-model.txt'
-		output = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-		gdown.download(url, output, quiet=False)
-		
-	ensemble_model_path = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-	
-	#print(ensemble_model_path)
-	
-	#-------------------------------
-	
-	deepface_ensemble = lgb.Booster(model_file = ensemble_model_path)
-
-	comp_pbar = tqdm(range(0, len(metrics_df.index)), desc = 'Realizando o ensemble para cada par de indivíduos!')
-
-	for index in comp_pbar:
-
-		i = metrics_df.index[index]
-				
-		prediction = deepface_ensemble.predict(np.expand_dims(np.array(metrics_df.loc[i]), axis=0))[0]
-	
-		verified = np.argmax(prediction) == 1
-		if verified: identified = "true"
-		else: identified = "false"
-	
-		score = prediction[np.argmax(prediction)]
-
-		names = i.split('-')
-
-		names = names[0].split('.') + names[1].split('.')
-
-		if names[0] == names[2]:
-			reality = "true"
-		else: reality = "false"
-
-		ensembles = [identified, reality, score]
-
-		verification_df.loc[i] = ensembles
-
-	return verification_df
-
-	#------------------------------
-
-def f1_calculation(verif_df):
-
-	TP, FP, TN, FN = 0, 0, 0, 0
-
-	column_names = ['Accuracy', 'Recall', 'Precision', 'F1']
-
-	for i in verif_df.index:
-
-		if (verif_df.loc[i][0] == verif_df.loc[i][1]) and (verif_df.loc[i][0] == 'true'):
-			TP += 1
-		elif (verif_df.loc[i][0] != verif_df.loc[i][1]) and (verif_df.loc[i][0] == 'true'):
-			FP += 1 
-		elif (verif_df.loc[i][0] == verif_df.loc[i][1]) and (verif_df.loc[i][0] != 'true'):
-			TN += 1
-		else:
-			FN += 1
-
-	accuracy = (TP+TN)/(TP+TN+FP+FN)
-
-	recall = TN/(TN+FN)
-
-	precision = TP/(TP+FP)
-
-	f1 = 2*(recall*precision)/(recall+precision)
-
-	f1_df = pd.DataFrame([list([accuracy, recall, precision, f1])], columns = column_names, index = ['Results:'])
-
-	return f1_df
-
-#------------------------------
-	
-def compare_img_df(img_path, img_label,dic,  df):
-
-	img_dic = {img_label: img_path}
-
-
-	df_upload = save_hash(img_dic)
-	df_upload.append(df, ignore_index=False)
-	frames = [df_upload, df]
-	df = pd.concat(frames) 	
-
-	metrics = metrics_dataframe(df, one_compair = True)
-	
-	ensemble = ensemble_dataframe(metrics)
-	ensemble = ensemble[ensemble['Verified'] == 'true']	
-	index = np.argmax(ensemble["Score"])
-	names = list(ensemble.index)[index]
-	print(names)
-	names = names.split("-")
-	names = names[0].split('.')
-	print(names)
-
-	return ensemble, names[0]		
-					
-	
-#------------------------------
-
-def verify(img1_path, img2_path=''
-	, model_name ='VGG-Face', distance_metric = 'cosine', model = None, enforce_detection = True):
-
-	tic = time.time()
-
-	if type(img1_path) == list:
-		bulkProcess = True
-		img_list = img1_path.copy()
-	else:
-		bulkProcess = False
-		img_list = [[img1_path, img2_path]]
-
-	#------------------------------
-	
-	resp_objects = []
-	
-	if model_name == 'Ensemble':
-		print("Ensemble learning enabled")
-		
-		import lightgbm as lgb #lightgbm==2.3.1
-		
-		if model == None:
-			model = {}
-			
-			model_pbar = tqdm(range(0, 4), desc='Face recognition models')
-			
-			for index in model_pbar:
-				
-				if index == 0:
-					model_pbar.set_description("Loading VGG-Face")
-					model["VGG-Face"] = VGGFace.loadModel()
-				elif index == 1:
-					model_pbar.set_description("Loading Google FaceNet")
-					model["Facenet"] = Facenet.loadModel()
-				elif index == 2:
-					model_pbar.set_description("Loading OpenFace")
-					model["OpenFace"] = OpenFace.loadModel()
-				elif index == 3:
-					model_pbar.set_description("Loading Facebook DeepFace")
-					model["DeepFace"] = FbDeepFace.loadModel()
-					
-		#--------------------------
-		#validate model dictionary because it might be passed from input as pre-trained
-		
-		found_models = []
-		for key, value in model.items():
-			found_models.append(key)
-		
-		if ('VGG-Face' in found_models) and ('Facenet' in found_models) and ('OpenFace' in found_models) and ('DeepFace' in found_models):
-			print("Ensemble learning will be applied for ", found_models," models")
-		else:
-			raise ValueError("You would like to apply ensemble learning and pass pre-built models but models must contain [VGG-Face, Facenet, OpenFace, DeepFace] but you passed "+found_models)
-			
-		#--------------------------
-		
-		model_names = ["VGG-Face", "Facenet", "OpenFace", "DeepFace"]
-		metrics = ["cosine", "euclidean", "euclidean_l2"]
-		
-		pbar = tqdm(range(0,len(img_list)), desc='Verification')
-		
-		#for instance in img_list:
-		for index in pbar:
-			instance = img_list[index]
-			
-			if type(instance) == list and len(instance) >= 2:
-				img1_path = instance[0]
-				img2_path = instance[1]
-				
-				ensemble_features = []; ensemble_features_string = "["
-				
-				for i in  model_names:
-					custom_model = model[i]
-					
-					#input_shape = custom_model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-					input_shape = custom_model.layers[0].input_shape
-					
-					if type(input_shape) == list:
-						input_shape = input_shape[0][1:3]
-					else:
-						input_shape = input_shape[1:3]
-					
-					
-					img1 = functions.detectFace(img1_path, input_shape, enforce_detection = enforce_detection)
-					img2 = functions.detectFace(img2_path, input_shape, enforce_detection = enforce_detection)
-					
-					img1_representation = custom_model.predict(img1)[0,:]
-					img2_representation = custom_model.predict(img2)[0,:]
-					
-					for j in metrics:
-						if j == 'cosine':
-							distance = dst.findCosineDistance(img1_representation, img2_representation)
-						elif j == 'euclidean':
-							distance = dst.findEuclideanDistance(img1_representation, img2_representation)
-						elif j == 'euclidean_l2':
-							distance = dst.findEuclideanDistance(dst.l2_normalize(img1_representation), dst.l2_normalize(img2_representation))
-						
-						if i == 'OpenFace' and j == 'euclidean': #this returns same with OpenFace - euclidean_l2
-							continue
-						else:
-							
-							ensemble_features.append(distance)
-							
-							if len(ensemble_features) > 1:
-								ensemble_features_string += ", "
-							ensemble_features_string += str(distance)
-							
-				#print("ensemble_features: ", ensemble_features)
-				ensemble_features_string += "]"
-				
-				#-------------------------------
-				#find deepface path
-				
-				home = str(Path.home())
-				
-				if os.path.isfile(home+'/.deepface/weights/face-recognition-ensemble-model.txt') != True:
-					print("face-recognition-ensemble-model.txt will be downloaded...")
-					url = 'https://raw.githubusercontent.com/serengil/deepface/master/deepface/models/face-recognition-ensemble-model.txt'
-					output = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-					gdown.download(url, output, quiet=False)
-					
-				ensemble_model_path = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-				
-				#print(ensemble_model_path)
-				
-				#-------------------------------
-				
-				deepface_ensemble = lgb.Booster(model_file = ensemble_model_path)
-				
-				prediction = deepface_ensemble.predict(np.expand_dims(np.array(ensemble_features), axis=0))[0]
-				
-				verified = np.argmax(prediction) == 1
-				if verified: identified = "true"
-				else: identified = "false"
-				
-				score = prediction[np.argmax(prediction)]
-				
-				#print("verified: ", verified,", score: ", score)
-				
-				resp_obj = "{"
-				resp_obj += "\"verified\": "+identified
-				resp_obj += ", \"score\": "+str(score)
-				resp_obj += ", \"distance\": "+ensemble_features_string
-				resp_obj += ", \"model\": [\"VGG-Face\", \"Facenet\", \"OpenFace\", \"DeepFace\"]"
-				resp_obj += ", \"similarity_metric\": [\"cosine\", \"euclidean\", \"euclidean_l2\"]"
-				resp_obj += "}"
-				
-				#print(resp_obj)
-				
-				resp_obj = json.loads(resp_obj) #string to json
-				
-				if bulkProcess == True:
-					resp_objects.append(resp_obj)
-				else:
-					return resp_obj
-				
-				#-------------------------------
-		
-		if bulkProcess == True:
-			resp_obj = "{"
-
-			for i in range(0, len(resp_objects)):
-				resp_item = json.dumps(resp_objects[i])
-
-				if i > 0:
-					resp_obj += ", "
-
-				resp_obj += "\"pair_"+str(i+1)+"\": "+resp_item
-			resp_obj += "}"
-			resp_obj = json.loads(resp_obj)
-			return resp_obj
-		
-		return None
-		
-	#ensemble learning block end
-	#--------------------------------
-	#ensemble learning disabled
-	
-	if model == None:
-		if model_name == 'VGG-Face':
-			print("Using VGG-Face model backend and", distance_metric,"distance.")
-			model = VGGFace.loadModel()
-
-		elif model_name == 'OpenFace':
-			print("Using OpenFace model backend", distance_metric,"distance.")
-			model = OpenFace.loadModel()
-
-		elif model_name == 'Facenet':
-			print("Using Facenet model backend", distance_metric,"distance.")
-			model = Facenet.loadModel()
-
-		elif model_name == 'DeepFace':
-			print("Using FB DeepFace model backend", distance_metric,"distance.")
-			model = FbDeepFace.loadModel()
-		
-		elif model_name == 'DeepID':
-			print("Using DeepID2 model backend", distance_metric,"distance.")
-			model = DeepID.loadModel()
-
-		else:
-			raise ValueError("Invalid model_name passed - ", model_name)
-	else: #model != None
-		print("Already built model is passed")
-
-	#------------------------------
-	#face recognition models have different size of inputs
-	#input_shape = model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-	input_shape = model.layers[0].input_shape
-	
-	if type(input_shape) == list:
-		input_shape = input_shape[0][1:3]
-	else:
-		input_shape = input_shape[1:3]
-  
-	input_shape_x = input_shape[0]
-	input_shape_y = input_shape[1]
-
-	#------------------------------
-
-	#tuned thresholds for model and metric pair
-	threshold = functions.findThreshold(model_name, distance_metric)
-
-	#------------------------------
-	pbar = tqdm(range(0,len(img_list)), desc='Verification')
-	
-	#for instance in img_list:
-	for index in pbar:
-	
-		instance = img_list[index]
-		
-		if type(instance) == list and len(instance) >= 2:
-			img1_path = instance[0]
-			img2_path = instance[1]
-
-			#----------------------
-			#crop and align faces
-
-			img1 = functions.detectFace(img1_path, (input_shape_y, input_shape_x), enforce_detection = enforce_detection)
-			img2 = functions.detectFace(img2_path, (input_shape_y, input_shape_x), enforce_detection = enforce_detection)
-
-			#----------------------
-			#find embeddings
-
-			img1_representation = model.predict(img1)[0,:]
-			img2_representation = model.predict(img2)[0,:]
-
-			#----------------------
-			#find distances between embeddings
-
-			if distance_metric == 'cosine':
-				distance = dst.findCosineDistance(img1_representation, img2_representation)
-			elif distance_metric == 'euclidean':
-				distance = dst.findEuclideanDistance(img1_representation, img2_representation)
-			elif distance_metric == 'euclidean_l2':
-				distance = dst.findEuclideanDistance(dst.l2_normalize(img1_representation), dst.l2_normalize(img2_representation))
-			else:
-				raise ValueError("Invalid distance_metric passed - ", distance_metric)
-
-			#----------------------
-			#decision
-
-			if distance <= threshold:
-				identified =  "true"
-			else:
-				identified =  "false"
-
-			#----------------------
-			#response object
-
-			resp_obj = "{"
-			resp_obj += "\"verified\": "+identified
-			resp_obj += ", \"distance\": "+str(distance)
-			resp_obj += ", \"max_threshold_to_verify\": "+str(threshold)
-			resp_obj += ", \"model\": \""+model_name+"\""
-			resp_obj += ", \"similarity_metric\": \""+distance_metric+"\""
-			resp_obj += "}"
-
-			resp_obj = json.loads(resp_obj) #string to json
-
-			if bulkProcess == True:
-				resp_objects.append(resp_obj)
-			else:
-				#K.clear_session()
-				return resp_obj
-			#----------------------
-
-		else:
-			raise ValueError("Invalid arguments passed to verify function: ", instance)
-
-	#-------------------------
-
-	toc = time.time()
-
-	#print("identification lasts ",toc-tic," seconds")
-
-	if bulkProcess == True:
-		resp_obj = "{"
-
-		for i in range(0, len(resp_objects)):
-			resp_item = json.dumps(resp_objects[i])
-
-			if i > 0:
-				resp_obj += ", "
-
-			resp_obj += "\"pair_"+str(i+1)+"\": "+resp_item
-		resp_obj += "}"
-		resp_obj = json.loads(resp_obj)
-		return resp_obj
-		#return resp_objects
-
-
-def analyze(img_path, actions = [], models = {}, enforce_detection = True):
-
-	if type(img_path) == list:
-		img_paths = img_path.copy()
-		bulkProcess = True
-	else:
-		img_paths = [img_path]
-		bulkProcess = False
-
-	#---------------------------------
-
-	#if a specific target is not passed, then find them all
-	if len(actions) == 0:
-		actions= ['emotion', 'age', 'gender', 'race']
-
-	print("Actions to do: ", actions)
-
-	#---------------------------------
-
-	if 'emotion' in actions:
-		if 'emotion' in models:
-			print("already built emotion model is passed")
-			emotion_model = models['emotion']
-		else:
-			emotion_model = Emotion.loadModel()
-
-	if 'age' in actions:
-		if 'age' in models:
-			print("already built age model is passed")
-			age_model = models['age']
-		else:
-			age_model = Age.loadModel()
-
-	if 'gender' in actions:
-		if 'gender' in models:
-			print("already built gender model is passed")
-			gender_model = models['gender']
-		else:
-			gender_model = Gender.loadModel()
-
-	if 'race' in actions:
-		if 'race' in models:
-			print("already built race model is passed")
-			race_model = models['race']
-		else:
-			race_model = Race.loadModel()
-	#---------------------------------
-
-	resp_objects = []
-	
-	global_pbar = tqdm(range(0,len(img_paths)), desc='Analyzing')
-	
-	#for img_path in img_paths:
-	for j in global_pbar:
-		img_path = img_paths[j]
-
-		resp_obj = "{"
-
-		#TO-DO: do this in parallel
-
-		pbar = tqdm(range(0,len(actions)), desc='Finding actions')
-
-		action_idx = 0
-		img_224 = None # Set to prevent re-detection
-		#for action in actions:
-		for index in pbar:
-			action = actions[index]
-			pbar.set_description("Action: %s" % (action))
-
-			if action_idx > 0:
-				resp_obj += ", "
-
-			if action == 'emotion':
-				emotion_labels = ['angry', 'disgust', 'fear', 'happy', 'sad', 'surprise', 'neutral']
-				img = functions.detectFace(img_path, target_size = (48, 48), grayscale = True, enforce_detection = enforce_detection)
-
-				emotion_predictions = emotion_model.predict(img)[0,:]
-
-				sum_of_predictions = emotion_predictions.sum()
-
-				emotion_obj = "\"emotion\": {"
-				for i in range(0, len(emotion_labels)):
-					emotion_label = emotion_labels[i]
-					emotion_prediction = 100 * emotion_predictions[i] / sum_of_predictions
-
-					if i > 0: emotion_obj += ", "
-
-					emotion_obj += "\"%s\": %s" % (emotion_label, emotion_prediction)
-
-				emotion_obj += "}"
-
-				emotion_obj += ", \"dominant_emotion\": \"%s\"" % (emotion_labels[np.argmax(emotion_predictions)])
-
-				resp_obj += emotion_obj
-
-			elif action == 'age':
-				if img_224 is None:
-					img_224 = functions.detectFace(img_path, target_size = (224, 224), grayscale = False, enforce_detection = enforce_detection) #just emotion model expects grayscale images
-				#print("age prediction")
-				age_predictions = age_model.predict(img_224)[0,:]
-				apparent_age = Age.findApparentAge(age_predictions)
-
-				resp_obj += "\"age\": %s" % (apparent_age)
-
-			elif action == 'gender':
-				if img_224 is None:
-					img_224 = functions.detectFace(img_path, target_size = (224, 224), grayscale = False, enforce_detection = enforce_detection) #just emotion model expects grayscale images
-				#print("gender prediction")
-
-				gender_prediction = gender_model.predict(img_224)[0,:]
-
-				if np.argmax(gender_prediction) == 0:
-					gender = "Woman"
-				elif np.argmax(gender_prediction) == 1:
-					gender = "Man"
-
-				resp_obj += "\"gender\": \"%s\"" % (gender)
-
-			elif action == 'race':
-				if img_224 is None:
-					img_224 = functions.detectFace(img_path, target_size = (224, 224), grayscale = False, enforce_detection = enforce_detection) #just emotion model expects grayscale images
-				race_predictions = race_model.predict(img_224)[0,:]
-				race_labels = ['asian', 'indian', 'black', 'white', 'middle eastern', 'latino hispanic']
-
-				sum_of_predictions = race_predictions.sum()
-
-				race_obj = "\"race\": {"
-				for i in range(0, len(race_labels)):
-					race_label = race_labels[i]
-					race_prediction = 100 * race_predictions[i] / sum_of_predictions
-
-					if i > 0: race_obj += ", "
-
-					race_obj += "\"%s\": %s" % (race_label, race_prediction)
-
-				race_obj += "}"
-				race_obj += ", \"dominant_race\": \"%s\"" % (race_labels[np.argmax(race_predictions)])
-
-				resp_obj += race_obj
-
-			action_idx = action_idx + 1
-
-		resp_obj += "}"
-
-		resp_obj = json.loads(resp_obj)
-
-		if bulkProcess == True:
-			resp_objects.append(resp_obj)
-		else:
-			return resp_obj
-
-	if bulkProcess == True:
-		resp_obj = "{"
-
-		for i in range(0, len(resp_objects)):
-			resp_item = json.dumps(resp_objects[i])
-
-			if i > 0:
-				resp_obj += ", "
-
-			resp_obj += "\"instance_"+str(i+1)+"\": "+resp_item
-		resp_obj += "}"
-		resp_obj = json.loads(resp_obj)
-		return resp_obj
-		#return resp_objects
-
-
-def detectFace(img_path):
-	img = functions.detectFace(img_path)[0] #detectFace returns (1, 224, 224, 3)
-	return img[:, :, ::-1] #bgr to rgb
-
-def find(img_path, db_path
-	, model_name ='VGG-Face', distance_metric = 'cosine', model = None, enforce_detection = True):
-		
-	tic = time.time()
-	
-	if type(img_path) == list:
-		bulkProcess = True
-		img_paths = img_path.copy()
-	else:
-		bulkProcess = False
-		img_paths = [img_path]
-	
-	if os.path.isdir(db_path) == True:
-		
-		#---------------------------------------
-		
-		if model == None:
-			if model_name == 'VGG-Face':
-				print("Using VGG-Face model backend and", distance_metric,"distance.")
-				model = VGGFace.loadModel()
-			elif model_name == 'OpenFace':
-				print("Using OpenFace model backend", distance_metric,"distance.")
-				model = OpenFace.loadModel()
-			elif model_name == 'Facenet':
-				print("Using Facenet model backend", distance_metric,"distance.")
-				model = Facenet.loadModel()
-			elif model_name == 'DeepFace':
-				print("Using FB DeepFace model backend", distance_metric,"distance.")
-				model = FbDeepFace.loadModel()
-			elif model_name == 'DeepID':
-				print("Using DeepID model backend", distance_metric,"distance.")
-				model = DeepID.loadModel()
-			elif model_name == 'Ensemble':
-				
-				print("Ensemble learning enabled")
-				#TODO: include DeepID in ensemble method
-				
-				import lightgbm as lgb #lightgbm==2.3.1
-				
-				model_names = ['VGG-Face', 'Facenet', 'OpenFace', 'DeepFace']
-				metric_names = ['cosine', 'euclidean', 'euclidean_l2']
-				models = {}
-				
-				pbar = tqdm(range(0, len(model_names)), desc='Face recognition models')
-				
-				for index in pbar:
-					if index == 0:
-						pbar.set_description("Loading VGG-Face")
-						models['VGG-Face'] = VGGFace.loadModel()
-					elif index == 1:
-						pbar.set_description("Loading FaceNet")
-						models['Facenet'] = Facenet.loadModel()
-					elif index == 2:
-						pbar.set_description("Loading OpenFace")
-						models['OpenFace'] = OpenFace.loadModel()
-					elif index == 3:
-						pbar.set_description("Loading DeepFace")
-						models['DeepFace'] = FbDeepFace.loadModel()
-						
-			else:
-				raise ValueError("Invalid model_name passed - ", model_name)	
-		else: #model != None
-			print("Already built model is passed")
-			
-			if model_name == 'Ensemble':
-				
-				#validate model dictionary because it might be passed from input as pre-trained
-				
-				found_models = []
-				for key, value in model.items():
-					found_models.append(key)
-				
-				if ('VGG-Face' in found_models) and ('Facenet' in found_models) and ('OpenFace' in found_models) and ('DeepFace' in found_models):
-					print("Ensemble learning will be applied for ", found_models," models")
-				else:
-					raise ValueError("You would like to apply ensemble learning and pass pre-built models but models must contain [VGG-Face, Facenet, OpenFace, DeepFace] but you passed "+found_models)
-		
-		#threshold = functions.findThreshold(model_name, distance_metric)
-		
-		#---------------------------------------
-		
-		file_name = "representations_%s.pkl" % (model_name)
-		file_name = file_name.replace("-", "_").lower()
-		
-		if path.exists(db_path+"/"+file_name):
-			
-			print("WARNING: Representations for images in ",db_path," folder were previously stored in ", file_name, ". If you added new instances after this file creation, then please delete this file and call find function again. It will create it again.")
-			
-			f = open(db_path+'/'+file_name, 'rb')
-			representations = pickle.load(f)
-			
-			print("There are ", len(representations)," representations found in ",file_name)
-			
-		else:
-			employees = []
-			
-			for r, d, f in os.walk(db_path): # r=root, d=directories, f = files
-				for file in f:
-					if ('.jpg' in file):
-						exact_path = r + "/" + file
-						employees.append(exact_path)
-			
-			if len(employees) == 0:
-				raise ValueError("There is no image in ", db_path," folder!")
-			
-			#------------------------
-			#find representations for db images
-			
-			representations = []
-			
-			pbar = tqdm(range(0,len(employees)), desc='Finding representations')
-			
-			#for employee in employees:
-			for index in pbar:
-				employee = employees[index]
-				
-				if model_name != 'Ensemble':
-				
-					#input_shape = model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-					input_shape = model.layers[0].input_shape
-					
-					if type(input_shape) == list:
-						input_shape = input_shape[0][1:3]
-					else:
-						input_shape = input_shape[1:3]
-
-					input_shape_x = input_shape[0]; input_shape_y = input_shape[1]
-					
-					img = functions.detectFace(employee, (input_shape_y, input_shape_x), enforce_detection = enforce_detection)
-					representation = model.predict(img)[0,:]
-					
-					instance = []
-					instance.append(employee)
-					instance.append(representation)
-					
-				else: #ensemble learning
-					
-					instance = []
-					instance.append(employee)
-					
-					for j in model_names:
-						model = models[j]
-						
-						#input_shape = model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-						input_shape = model.layers[0].input_shape
-						
-						if type(input_shape) == list:
-							input_shape = input_shape[0][1:3]
-						else:
-							input_shape = input_shape[1:3]
-						
-						input_shape_x = input_shape[0]; input_shape_y = input_shape[1]
-						
-						img = functions.detectFace(employee, (input_shape_y, input_shape_x), enforce_detection = enforce_detection)
-						representation = model.predict(img)[0,:]
-						instance.append(representation)
-				
-				#-------------------------------
-				
-				representations.append(instance)
-			
-			f = open(db_path+'/'+file_name, "wb")
-			pickle.dump(representations, f)
-			f.close()
-			
-			print("Representations stored in ",db_path,"/",file_name," file. Please delete this file when you add new identities in your database.")
-		
-		#----------------------------
-		#we got representations for database
-		
-		if model_name != 'Ensemble':
-			df = pd.DataFrame(representations, columns = ["identity", "representation"])
-		else: #ensemble learning
-			df = pd.DataFrame(representations, columns = ["identity", "VGG-Face_representation", "Facenet_representation", "OpenFace_representation", "DeepFace_representation"])
-			
-		df_base = df.copy()
-		
-		resp_obj = []
-		
-		global_pbar = tqdm(range(0,len(img_paths)), desc='Analyzing')
-		for j in global_pbar:
-			img_path = img_paths[j]
-		
-			#find representation for passed image
-			
-			if model_name == 'Ensemble':
-				for j in model_names:
-					model = models[j]
-					
-					#input_shape = model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-					input_shape = model.layers[0].input_shape
-					
-					if type(input_shape) == list:
-						input_shape = input_shape[0][1:3]
-					else:
-						input_shape = input_shape[1:3]
-					
-					img = functions.detectFace(img_path, input_shape, enforce_detection = enforce_detection)
-					target_representation = model.predict(img)[0,:]
-					
-					for k in metric_names:
-						distances = []
-						for index, instance in df.iterrows():
-							source_representation = instance["%s_representation" % (j)]
-							
-							if k == 'cosine':
-								distance = dst.findCosineDistance(source_representation, target_representation)
-							elif k == 'euclidean':
-								distance = dst.findEuclideanDistance(source_representation, target_representation)
-							elif k == 'euclidean_l2':
-								distance = dst.findEuclideanDistance(dst.l2_normalize(source_representation), dst.l2_normalize(target_representation))
-							
-							distances.append(distance)
-						
-						if j == 'OpenFace' and k == 'euclidean':
-							continue
-						else:
-							df["%s_%s" % (j, k)] = distances
-				
-				#----------------------------------
-				
-				feature_names = []
-				for j in model_names:
-					for k in metric_names:
-						if j == 'OpenFace' and k == 'euclidean':
-							continue
-						else:
-							feature = '%s_%s' % (j, k)
-							feature_names.append(feature)
-				
-				#print(df[feature_names].head())
-				
-				x = df[feature_names].values
-				
-				#----------------------------------
-				#lightgbm model
-				home = str(Path.home())
-				
-				if os.path.isfile(home+'/.deepface/weights/face-recognition-ensemble-model.txt') != True:
-					print("face-recognition-ensemble-model.txt will be downloaded...")
-					url = 'https://raw.githubusercontent.com/serengil/deepface/master/deepface/models/face-recognition-ensemble-model.txt'
-					output = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-					gdown.download(url, output, quiet=False)
-					
-				ensemble_model_path = home+'/.deepface/weights/face-recognition-ensemble-model.txt'
-				
-				deepface_ensemble = lgb.Booster(model_file = ensemble_model_path)
-				
-				y = deepface_ensemble.predict(x)
-				
-				verified_labels = []; scores = []
-				for i in y:
-					verified = np.argmax(i) == 1
-					score = i[np.argmax(i)]
-					
-					verified_labels.append(verified)
-					scores.append(score)
-				
-				df['verified'] = verified_labels
-				df['score'] = scores
-				
-				df = df[df.verified == True]
-				#df = df[df.score > 0.99] #confidence score
-				df = df.sort_values(by = ["score"], ascending=False).reset_index(drop=True)
-				df = df[['identity', 'verified', 'score']]
-				
-				resp_obj.append(df)
-				df = df_base.copy() #restore df for the next iteration
-				
-				#----------------------------------
-			
-			if model_name != 'Ensemble':
-			
-				#input_shape = model.layers[0].input_shape[1:3] #my environment returns (None, 224, 224, 3) but some people mentioned that they got [(None, 224, 224, 3)]. I think this is because of version issue.
-	
-				input_shape = model.layers[0].input_shape
-				
-				if type(input_shape) == list:
-					input_shape = input_shape[0][1:3]
-				else:
-					input_shape = input_shape[1:3]
-				
-				input_shape_x = input_shape[0]; input_shape_y = input_shape[1]
-				
-				img = functions.detectFace(img_path, (input_shape_y, input_shape_x), enforce_detection = enforce_detection)
-				target_representation = model.predict(img)[0,:]
-		
-				distances = []
-				for index, instance in df.iterrows():
-					source_representation = instance["representation"]
-					
-					if distance_metric == 'cosine':
-						distance = dst.findCosineDistance(source_representation, target_representation)
-					elif distance_metric == 'euclidean':
-						distance = dst.findEuclideanDistance(source_representation, target_representation)
-					elif distance_metric == 'euclidean_l2':
-						distance = dst.findEuclideanDistance(dst.l2_normalize(source_representation), dst.l2_normalize(target_representation))
-					else:
-						raise ValueError("Invalid distance_metric passed - ", distance_metric)
-					
-					distances.append(distance)
-				
-				threshold = functions.findThreshold(model_name, distance_metric)
-				
-				df["distance"] = distances
-				df = df.drop(columns = ["representation"])
-				df = df[df.distance <= threshold]
-			
-				df = df.sort_values(by = ["distance"], ascending=True).reset_index(drop=True)
-				resp_obj.append(df)
-				df = df_base.copy() #restore df for the next iteration
-			
-		toc = time.time()
-		
-		print("find function lasts ",toc-tic," seconds")
-		
-		if len(resp_obj) == 1:
-			return resp_obj[0]
-		
-		return resp_obj
-		
-	else:
-		raise ValueError("Passed db_path does not exist!")
-		
-	return None
-	
-def stream(db_path = '', model_name ='VGG-Face', distance_metric = 'cosine', enable_face_analysis = True):
-	realtime.analysis(db_path, model_name, distance_metric, enable_face_analysis)
-
-def allocateMemory():
-	print("Analyzing your system...")
-	functions.allocateMemory()
-
-functions.initializeFolder()
-
-#---------------------------
