+++ apresentacao_do_projeto/.vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "/home/fill/anaconda3/envs/landmarks/bin/python"
+}
