+++ códigos/vgg/README.md
@@ -1,0 +1,9 @@
+# Reconhecimento de Rostos com VGG
+
+* [Rede utilizada](https://drive.google.com/file/d/1CPSeum3HpopfomUEK1gybeuIVoeJT_Eo/view "Rede Utilizada")
+
+* [Dataset Utilizado]()
+
+Esse código utiliza a rede para calcular um vetor de dimensão 2622 com base no rosto de pessoas. Com isso é possivel utilizar uma distância de cossenos para separar pessoas diferentes.
+
+(Colocar imagens legais aqui depois)
